backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Neo4j driver
neo4j==5.17.0

# HTTP requests for Solr integration
requests==2.31.0

# Additional utilities
python-dotenv==1.0.0
jinja2==3.1.2

# Optional: For future LLM integration
# torch>=2.0.0
# transformers>=4.35.0
# accelerate>=0.24.0